+++ test/shared-tests/with-union-type-fragment-spread.ts
@@ -1,0 +1,52 @@
+import gql from "graphql-tag";
+import { SharedTestDef } from "../shared-test-def";
+
+export const test: SharedTestDef = {
+  // only: true,
+  name: "with union type fragment spread",
+  query: gql`
+    query TestQuery {
+      booksAndAuthors {
+        id
+        __typename
+        ...BookFragment
+        ...AuthorFragment
+      }
+    }
+    fragment BookFragment on Book {
+      title
+    }
+    fragment AuthorFragment on Author {
+      name
+    }
+  `,
+  data: {
+    booksAndAuthors: [
+      {
+        id: "123",
+        __typename: "Book",
+        title: "My awesome blog post"
+      },
+      {
+        id: "324",
+        __typename: "Author",
+        name: "Nicole"
+      }
+    ]
+  },
+  normMap: {
+    ROOT_QUERY: {
+      booksAndAuthors: ["Book:123", "Author:324"]
+    },
+    "Book:123": {
+      id: "123",
+      __typename: "Book",
+      title: "My awesome blog post"
+    },
+    "Author:324": {
+      id: "324",
+      __typename: "Author",
+      name: "Nicole"
+    }
+  }
+};